AI_agent_expert_system/backend/requirements.txt
# 現有套件
streamlit>=1.30.0
python-pptx>=0.6.23
Pillow>=10.0.0
openai>=1.10.0
python-dotenv>=1.0.0
tenacity>=8.2.0
pandas>=2.0.0
httpx>=0.25.0
sqlite-vec>=0.1.0
pymupdf>=1.23.0
rapidfuzz>=3.0.0

# FastAPI 相關
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# 檔案監控
watchdog>=4.0.0

# Gemini SDK
google-generativeai>=0.3.0

# 開發工具
pytest>=8.0.0
pytest-asyncio>=0.23.0
